deployment/requirements.txt
Flask==2.1.0
Pillow
numpy
